# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pearlis Luxury Jewelry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders (or all orders for admin)
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # USERS / AUTH
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons (admin)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # NEWSLETTER
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeBody"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # DASHBOARD (admin)
...
  /dashboard/recent-orders:
    get:
      operationId: getDashboardRecentOrders
      tags: [dashboard]
      summary: Get recent orders for dashboard
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /dashboard/top-products:
    get:
      operationId: getDashboardTopProducts
      tags: [dashboard]
      summary: Get top-selling products
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    Product:
      type: object
      required: [id, name, price, category, images, stock, isNew, isTrending, rating, reviewCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
          nullable: true
        category:
          type: string
        categoryId:
          type: integer
        material:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isNew:
          type: boolean
        isTrending:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    OrderItem:
      type: object
      required: [productId, quantity, price, productName, productImage]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        productName:
          type: string
        productImage:
          type: string
    Order:
      type: object
      required: [id, status, total, items, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        total:
          type: number
        subtotal:
          type: number
        discount:
          type: number
        couponCode:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        shippingAddress:
          $ref: "#/components/schemas/Address"
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        createdAt:
          type: string
        customerName:
          type: string
          nullable: true
        customerEmail:
          type: string
          nullable: true
    OrderListResponse:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CreateOrderBody:
      type: object
      required: [shippingAddress, paymentMethod]
      properties:
        shippingAddress:
          $ref: "#/components/schemas/CreateAddressBody"
        paymentMethod:
          type: string
        couponCode:
          type: string
          nullable: true
    UpdateOrderStatusBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
...
    Address:
      type: object
      required: [id, name, line1, city, state, postalCode, country]
      properties:
        id:
          type: integer
        name:
          type: string
        line1:
          type: string
        line2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
        phone:
          type: string
          nullable: true
        isDefault:
          type: boolean
    CreateAddressBody:
      type: object
      required: [name, line1, city, state, postalCode, country]
      properties:
        name:
          type: string
        line1:
          type: string
        line2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
        phone:
          type: string
          nullable: true
        isDefault:
          type: boolean
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue, isActive]
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
          nullable: true
        maxUses:
          type: integer
          nullable: true
        usedCount:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true
    CreateCouponBody:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
          nullable: true
        maxUses:
          type: integer
          nullable: true
        expiresAt:
          type: string
          nullable: true
    NewsletterSubscribeBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
...